[Restore] fix sql to sql restoring

+++ Account/Controller/controllerdb.cpp
@@ -24,23 +24,19 @@ ControllerDB::~ControllerDB()
 
 bool ControllerDB::init()
 {
-    if(m_db.connectionNames().contains("default"))
+    if(!backup && m_db.connectionNames().contains("default"))
     {
         m_db.database("default").close();
         m_db.removeDatabase("default");
     }
     
-    if(m_db.connectionNames().contains("backup"))
+    if(backup && m_db.connectionNames().contains("backup"))
     {
         m_db.database("backup").close();
         m_db.removeDatabase("backup");
     }
 
-    if(backup)
-    {
-        QDir dir;
-        dir.remove("account_backup");
-    }
+
     m_db = QSqlDatabase::addDatabase("QSQLITE", backup ? "backup" : "default");
     qDebug()<<"Back?"<<backup<<m_db.connectionNames()<<m_db.connectionName();
     QString name = QString("account%1").arg(backup ? "_backup" : "");
